Remove time stepping within coupling interval
Remove time stepping within the the time interval which led to physical inconsistency if a segment died before the end of a coupling interval and was thus not evolved and advected for the whole interval duration.

Reorder main steps so advect comes before dump. This corrects the physical inconsistency that dumped segments were not advected through the coupling interval.

+++ src/timekeeping.h
@@ -131,7 +131,7 @@ public:
 
 // Non-member operators for CMTimeInterval
 
-inline CMTimeInterval operator*(const CMTimeInterval& timeInt, float scalar) {
+inline CMTimeInterval operator*(const CMTimeInterval& timeInt, double scalar) {
     CMTimeInterval newTimeInt;
     newTimeInt.set(timeInt.yy*scalar,
                    timeInt.mm*scalar,
@@ -142,7 +142,7 @@ inline CMTimeInterval operator*(const CMTimeInterval& timeInt, float scalar) {
     return newTimeInt;
 }
 
-inline CMTimeInterval operator*(float scalar, const CMTimeInterval& timeInt) {
+inline CMTimeInterval operator*(double scalar, const CMTimeInterval& timeInt) {
     return timeInt*scalar;
 }
